[Bronze III] Title: 수학은 체육과목 입니다, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/15894. 수학은 체육과목 입니다/수학은 체육과목 입니다.cc
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(void) {
-    long long n;
+    long n;
     cin >> n;
     cout << n * 4 << endl;
     return 0;
